+++ src/toBeSorted/unk_flag_stuff.cpp
@@ -24,7 +24,7 @@ u16 UnkFlagStuff::maskForIdx(u16 index) {
     return calculateMask(mpDefinitions[index].mShiftMask);
 }
 
-u32 UnkFlagStuff::getCounterOrFlag(u16 counterIdx, u16 *dataPtr, u32 param4) {
+u16 UnkFlagStuff::getCounterOrFlag(u16 counterIdx, u16 *dataPtr, u32 param4) {
     u16 index = 0;
     u8 shift = 0;
     u16 mask = 0;
